Move visualization to different file
